no private types in public interfaces; backwards compatibility
